Move HttpClient related files to http subfolder
