+++ snowplow-cpp-tracker.vcxproj.filters
@@ -24,10 +24,10 @@
     <ClCompile Include="src\emitter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\http_client_windows.cpp">
+    <ClCompile Include="src\http\http_client_windows.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\http_request_result.cpp">
+    <ClCompile Include="src\http\http_request_result.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="src\payload.cpp">
@@ -54,7 +54,7 @@
     <ClCompile Include="include\sqlite3.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="test\test_http_client.cpp">
+    <ClCompile Include="test\http\test_http_client.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="test\client_session_test.cpp">
@@ -66,10 +66,10 @@
     <ClCompile Include="test\emitter_test.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="test\http_client_test.cpp">
+    <ClCompile Include="test\http\http_client_test.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="test\http_request_result_test.cpp">
+    <ClCompile Include="test\http\http_request_result_test.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="test\main.cpp">
@@ -107,13 +107,13 @@
     <ClInclude Include="src\emitter.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\http_client.hpp">
+    <ClInclude Include="src\http\http_client.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\http_client_windows.hpp">
+    <ClInclude Include="src\http\http_client_windows.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\http_request_result.hpp">
+    <ClInclude Include="src\http\http_request_result.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\payload.hpp">
@@ -146,7 +146,7 @@
     <clinclude include="test\catch.hpp">
       <filter>header files</filter>
     </clinclude>
-    <clinclude include="test\test_http_client.hpp">
+    <clinclude include="test\http\test_http_client.hpp">
       <filter>header files</filter>
     </clinclude>
   </ItemGroup>
