+++ test/emitter_test.cpp
@@ -12,7 +12,7 @@ See the Apache License Version 2.0 for the specific language governing permissio
 */
 
 #include "../src/emitter.hpp"
-#include "test_http_client.hpp"
+#include "http/test_http_client.hpp"
 #include "catch.hpp"
 
 using namespace snowplow;
